+++ src/www/api.ts
@@ -1,8 +1,9 @@
 
 // Core API
 export {
-    Insets,
-    IInsetsAPI,
-    IInsets,
-    IInsetCallbackFunc
-} from './Insets';
+    Inset
+} from './Inset';
+export {IInset} from './IInset';
+export {InsetMask} from './InsetMask';
+export {IInsetCallbackFunc} from './IInsetCallbackFunc';
+export {IInsetConfiguration} from './IInsetConfiguration';